+++ src/Replication/Memcache.php
@@ -2,6 +2,9 @@
 
 namespace Phlib\DbHelperReplication\Replication;
 
+use Phlib\DbHelperReplication\Exception\InvalidArgumentException;
+use Phlib\DbHelperReplication\Exception\RuntimeException;
+
 /**
  * @package Phlib\DbHelperReplication
  * @licence LGPL-3.0
@@ -9,27 +12,39 @@ namespace Phlib\DbHelperReplication\Replication;
 class Memcache implements StorageInterface
 {
     /**
-     * @var Memcache
-     */
-    protected $memcache;
-
-    /**
-     * @param \Memcache $memcache
-     */
-    public function __construct(\Memcache $memcache)
-    {
-        $this->memcache = $memcache;
-    }
-
-    /**
      * @param array $memcacheConfig
      * @return static
      */
     public static function createFromConfig(array $memcacheConfig)
     {
-        $memcache = new \Memcache();
-        $memcache->connect($memcacheConfig['host'], $memcacheConfig['port'], $memcacheConfig['timeout']);
+        $memcache = new \Memcached();
+
+        if (!isset($memcacheConfig['host'])) {
+            throw new InvalidArgumentException('Missing Memcache host');
+        }
+
+        $host = $memcacheConfig['host'];
+        $port = $memcacheConfig['port'] ?? 11211;
+        $timeout = $memcacheConfig['timeout'] ?? 200;
+
+        $memcache->setOption(\Memcached::OPT_CONNECT_TIMEOUT, $timeout); // ms
+        if ($memcache->addServer($host, $port) === false) {
+            throw new RuntimeException('Unable to connect to Memcache');
+        }
         return new static($memcache);
+    }
+
+    /**
+     * @var \Memcached
+     */
+    protected $memcache;
+
+    /**
+     * @param \Memcached $memcache
+     */
+    public function __construct(\Memcached $memcache)
+    {
+        $this->memcache = $memcache;
     }
 
     /**
@@ -56,7 +71,12 @@ class Memcache implements StorageInterface
     public function setSecondsBehind($host, $value)
     {
         $key = $this->getKey($host) . ':SecondsBehind';
-        return $this->memcache->set($key, (int)$value);
+        $result = $this->memcache->set($key, (int)$value);
+        if ($result === false) {
+            throw new RuntimeException('Unable to store value to Memcache');
+        }
+
+        return $this;
     }
 
     /**
@@ -74,6 +94,11 @@ class Memcache implements StorageInterface
     public function setHistory($host, array $values)
     {
         $key = $this->getKey($host) . ':History';
-        return $this->memcache->set($key, serialize($values));
+        $result = $this->memcache->set($key, serialize($values));
+        if ($result === false) {
+            throw new RuntimeException('Unable to store value to Memcache');
+        }
+
+        return $this;
     }
 }