+++ CHANGELOG.md
@@ -16,7 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Reduce visibility of internal methods and properties.
 ### Removed
 - Removed `replication:stats` CLI command which was not functional.
-- Removed support for PHP versions < v7.1 as they are no longer
+- Removed support for PHP versions <= v7.3 as they are no longer
   [actively supported](https://php.net/supported-versions.php) by the PHP project.
 
 ## [0.2.0] - 2021-09-22