+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- Add support for PHP v8.
 ### Fixed
 - `replication:monitor` CLI command did not initialise Replication when running
   directly, without `--daemonize`.